% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{estimate_migration_from_trajectories_C}
\alias{estimate_migration_from_trajectories_C}
\title{Internal C function (do not use).}
\usage{
estimate_migration_from_trajectories_C(
  migration,
  trajectories,
  estimate_transitions = NULL,
  estimate_sojourns = NULL,
  sojourn_domain = NULL,
  min_prob = NULL
)
}
\arguments{
\item{migration}{A migration (migration structure).}

\item{trajectories}{Set of trajectories (trajectory structure).}

\item{estimate_transitions}{Boolean for requiring transition parameters estimation.}

\item{estimate_sojourns}{Boolean for requiring sojourn parameters estimation.}

\item{sojourn_domain}{Sojourn parameters domains.}

\item{min_prob}{Minimum probability.}
}
\value{
migration with estimated parameters (migration structure).
}
\description{
Estimate parameters from a set of trajectories
}
