% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{estimate_migration_MCEM_C}
\alias{estimate_migration_MCEM_C}
\title{Internal C function (do not use).}
\usage{
estimate_migration_MCEM_C(
  migration,
  estimate_transitions = NULL,
  estimate_sojourns = NULL,
  sojourn_domain = NULL,
  start_transitions = NULL,
  start_sojourns = NULL,
  nb_particles = NULL,
  itermax = NULL,
  MC_algo = NULL,
  MH_neighborhood = NULL,
  MH_transition_length = NULL,
  log_transitions = NULL,
  log_sojourns = NULL,
  log_loglikelihood = NULL,
  log_sel_particles = NULL,
  log_acceptance_rate = NULL,
  verbose = NULL
)
}
\arguments{
\item{migration}{A migration (migration structure).}

\item{estimate_transitions}{Boolean for requiring transition parameters estimation.}

\item{estimate_sojourns}{Boolean for requiring sojourn parameters estimation.}

\item{sojourn_domain}{Sojourn parameters domains.}

\item{start_transitions}{Start transition.}

\item{start_sojourns}{Start sojourns.}

\item{nb_particles}{Number of particles.}

\item{itermax}{Maximum number of iteration.}

\item{MC_algo}{Monte Carlo algorithm name.}

\item{MH_neighborhood}{Neighborhood, only when MC_algo is "MH".}

\item{MH_transition_length}{Transition length, only when MC_algo is "MH".}

\item{log_transitions}{Log of transition parameters.}

\item{log_sojourns}{Log of sojourn parameters.}

\item{log_loglikelihood}{log of log-likelihood.}

\item{log_sel_particles}{log of selected particles.}

\item{log_acceptance_rate}{log of acceptance rate.}

\item{verbose}{Boolean for verbose.}
}
\value{
migration with estimated parameters (migration structure).
}
\description{
Estimate parameters with MCEM algorithm
}
