% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_toy_migration.R
\name{generate_toy_migration}
\alias{generate_toy_migration}
\title{Generate a toy migration structure with 5 sites.}
\usage{
generate_toy_migration()
}
\value{
A migration structure.
}
\description{
Generate a toy migration structure with 5 sites, for inference or simulation.
}
\examples{
toy_migration <- generate_toy_migration()
}
