% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{generate_trajectories_C}
\alias{generate_trajectories_C}
\title{Internal C function (do not use).}
\usage{
generate_trajectories_C(migration, trajectory = NULL, end_time = NULL)
}
\arguments{
\item{migration}{A migration (migration structure).}

\item{trajectory}{Set of trajectories partially simulated (trajectory structure).
Default simulated until time step 1 (initial state of each bird).}

\item{end_time}{Time of the end of simulation. Default is the horizon (last time step).}
}
\value{
the simulated trajectories
}
\description{
Simulates trajectories of individual birds.
}
