% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_counts_C}
\alias{get_counts_C}
\title{Internal C function (do not use).}
\usage{
get_counts_C(migration, trajs)
}
\arguments{
\item{migration}{A migration (migration structure).}

\item{trajs}{Set of trajectories (trajectory matrix).}
}
\value{
counts (count matrix).
}
\description{
Counts birds on each state for every time step.
}
