% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_observation_probability_C}
\alias{get_observation_probability_C}
\title{Internal C function (do not use).}
\usage{
get_observation_probability_C(obs, hidden_count, use_log)
}
\arguments{
\item{obs}{Observations. Matrix of I+2 rows and T+1 columns with observed 
bird counts. I is the numbers of sites, to which one adds the virtual 
sites 'death' and 'flight' and T is the horizon of the migration. 
structure.}

\item{hidden_count}{Hidden counts. Matrix of I+2 rows and T+1 columns with 
simulated bird counts. As computed by function get_count.}

\item{use_log}{Boolean that states if the log probability should be returned.}
}
\value{
the probability (or log-probability) of the observed data given 
  simulated counts.
}
\description{
Computes the probability of observations given the birds count from 
trajectories : P(O | Pi_[0:T])
}
