% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_paramlikelihood.R
\name{get_paramlikelihood}
\alias{get_paramlikelihood}
\title{Computes, by simulation, the log-likelihood of parameters}
\usage{
get_paramlikelihood(migr, obs, n_simul = 10)
}
\arguments{
\item{migr}{the migration structure which parameters likelihood 
should be evaluated}

\item{obs}{observation matrix used for the likelihood evaluation}

\item{n_simul}{number of simulations}
}
\value{
the log likelihood of the parameters of the migration structure migr.
}
\description{
Computes, by simulation, the log-likelihood of parameters   
              P(O | Lambda)
}
\examples{
migr <- generate_toy_migration()
get_paramlikelihood( migr, migr$observation)
}
