% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.migration.R
\name{plot.migration}
\alias{plot.migration}
\title{Plot a migration network from transition probabilities and mean sojourn time parameters.}
\usage{
\method{plot}{migration}(x, ...)
}
\arguments{
\item{x}{a migration structure}

\item{...}{others default arguments}
}
\value{
An igraph structure.
}
\description{
Plot a migration network from transition probabilities 
              and mean sojourn time parameters. 
              Currently just plot transition_law_param as edge width
             and sejourn_law_param as node size of the network
}
\examples{
migr <- generate_toy_migration()
plot( migr )
}
