% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_observedcounts.R
\name{plot_observedcounts}
\alias{plot_observedcounts}
\title{Barplot observed counts at different sites and time steps.}
\usage{
plot_observedcounts(migr, obs)
}
\arguments{
\item{migr}{A migration structure.}

\item{obs}{An observation matrix (S+2) x (H +1).}
}
\value{
No return value.
}
\description{
Barplot observed counts at different sites and time steps.
}
\examples{
migr <- generate_toy_migration()
traj <- generate_trajectories( migr )
obs <- generate_observedcounts( migr, traj )
plot_observedcounts( migr,  obs )
}
