% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_migration.R
\name{write_migration}
\alias{write_migration}
\title{Write a migration structure in a text file.}
\usage{
write_migration(migr, file_name)
}
\arguments{
\item{migr}{A migration structure.}

\item{file_name}{A file name.}
}
\value{
No returned value.
}
\description{
Write a migration structure in a text file.
If file exists, just print a warning and erase file.
}
\examples{
migr <- generate_toy_migration()
write_migration( migr, "toy_migration.txt" )
unlink("toy_migration.txt") # delete the file to pass package check
}
